Move tests to one class
