+++ collect_app/src/androidTest/java/org/odk/collect/android/feature/formentry/InstanceSerializationTest.kt
@@ -9,18 +9,65 @@ import org.junit.Test
 import org.junit.rules.RuleChain
 import org.junit.runner.RunWith
 import org.odk.collect.android.support.TestDependencies
+import org.odk.collect.android.support.pages.FormHierarchyPage
 import org.odk.collect.android.support.pages.MainMenuPage
 import org.odk.collect.android.support.rules.CollectTestRule
+import org.odk.collect.android.support.rules.RecentAppsRule
 import org.odk.collect.android.support.rules.TestRuleChain
 
 @RunWith(AndroidJUnit4::class)
 class InstanceSerializationTest {
     private val rule = CollectTestRule(useDemoProject = false)
+    private val recentAppsRule = RecentAppsRule()
     private val testDependencies = TestDependencies()
 
     @get:Rule
     val ruleChain: RuleChain = TestRuleChain.chain(testDependencies)
+        .around(recentAppsRule)
         .around(rule)
+
+    @Test
+    fun savingDraft_doesNotPruneNonRelevantNodes() {
+        testDependencies.server.addForm("one-question-relevance.xml")
+
+        rule.withProject(testDependencies.server.url, matchExactly = true)
+            .startBlankForm("One Question Relevance")
+            .clickOnText("Yes")
+            .swipeToNextQuestion("what is your age")
+            .answerQuestion("what is your age", "30")
+            .swipeToPreviousQuestion("Do you want to continue?")
+            .clickOnText("No")
+            .pressBackAndSaveAsDraft()
+            .clickDrafts(1)
+            .clickOnForm("One Question Relevance")
+            .clickOnQuestion("Do you want to continue?")
+            .clickOnText("Yes")
+            .swipeToNextQuestion("what is your age")
+            .assertAnswer("what is your age", "30")
+    }
+
+    @Test
+    fun savepoint_doesNotPruneNonRelevantNodes() {
+        testDependencies.server.addForm("one-question-relevance.xml")
+
+        rule.withProject(testDependencies.server.url, matchExactly = true)
+            .startBlankForm("One Question Relevance")
+            .clickOnText("Yes")
+            .swipeToNextQuestion("what is your age")
+            .answerQuestion("what is your age", "30")
+            .swipeToPreviousQuestion("Do you want to continue?")
+            .clickOnText("No")
+
+        recentAppsRule.leaveAndKillApp()
+
+        rule.reopenApp()
+            .startBlankFormWithSavepoint("One Question Relevance")
+            .clickRecover(FormHierarchyPage("One Question Relevance"))
+            .clickOnQuestion("Do you want to continue?")
+            .clickOnText("Yes")
+            .swipeToNextQuestion("what is your age")
+            .assertAnswer("what is your age", "30")
+    }
 
     @Test
     fun finalizingForm_doesPrunesNonRelevantNodes() {
